requirements.txt
# Vision-Language Model APIs
anthropic==0.8.1
openai==1.6.1
google-generativeai==0.3.2

# Core ML Libraries
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.4
scikit-image==0.22.0

# API and Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# HTTP Requests for DVLA API
requests==2.31.0
httpx==0.25.2

# Data Processing
pandas==2.1.4
python-dateutil==2.8.2

# Configuration and Environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging and Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Additional ML/AI Libraries
huggingface-hub==0.19.4
datasets==2.15.0

# Image format support
imageio==2.33.1

# File type detection
python-magic==0.4.27

# Async file operations
aiofiles==23.2.0
